Changing stop flag to uint64_t

+++ host/include/uhd/transport/async_recv_manager.hpp
@@ -166,7 +166,7 @@ private:
     std::thread* recv_loops;
 
     // Flag used to tell receive threads when to stop
-    uint_fast8_t stop_flag = false;
+    uint64_t stop_flag = false;
 
 public:
